small refactoring and implement requested changes
